Move `getCronNextRun` to a `useCron` hook
